Simplified Makefile structure for ST example
Included the whole HAL in compiled files, since the contents of the
unnecessary files are removed by the pre-processor anyway, based on the
HAL configuration header file.
The main STM32F4-Discovery BSP file has unfortunately no corresponding
mechanism, so one is forced to included many HAL components that are
not necessarily used in the application (SPI, etc.).