Fixed Vulkan ld path issue
In Ubuntu since a while back, LD_LIBRARY_PATH cannot be set in
.profile for security reason. The right way is to add a configuration
file under /etc/ld.so.conf.d/ instead. And to not forget to run
ldconfig!

This is particulary confusing in the case of the CMake configuration I
am using, where libvulkan.so is found anyway, probably because of some
magic during the building of glfw (with the apparent drawback of
setting an explicit lib path in the binary - the RUNPATH). That
explicit lib path, however, seems not to work for dlopen(), which is
used by the Vulkan SDK for validation. I have sweated quite a bit on
that one...